+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com) and this project follows the [Semantic Versioning](http://semver.org) code.
 
+## 0.4.0 - 2017-05-05
+### Added
+- Touch version of mobile-debug library
+
+### Changed
+- Build tasks
+
 ## 0.3.0 - 2017-04-25
 ### Changed
 - Global dependencies are now included